app/requirements.txt
numpy==1.18.4
Flask==1.1.2
pybase64==1.0.1
opencv-python==4.2.0.32
jsonpickle==1.4.1
pillow==7.1.2
gunicorn==20.0.4
ffmpeg-python
flask-cors